[d3d8] Refactor D3D8SwapChain implementation
