+++ src/d3d8/d3d8_swapchain.h
@@ -2,7 +2,6 @@
 
 #include "d3d8_device_child.h"
 #include "d3d8_surface.h"
-#include "d3d8_d3d9_util.h"
 
 namespace dxvk {
   
@@ -12,39 +11,19 @@ namespace dxvk {
   public:
 
     D3D8SwapChain(
-      D3D8Device*                      pDevice,
-      D3DPRESENT_PARAMETERS*           pPresentationParameters,
-      Com<d3d9::IDirect3DSwapChain9>&& pSwapChain)
-      : D3D8SwapChainBase(pDevice, std::move(pSwapChain)) {
-        m_backBuffers.resize(pPresentationParameters->BackBufferCount);
-    }
+            D3D8Device*                      pDevice,
+            D3DPRESENT_PARAMETERS*           pPresentationParameters,
+            Com<d3d9::IDirect3DSwapChain9>&& pSwapChain);
 
-    HRESULT STDMETHODCALLTYPE Present(const RECT *src, const RECT *dst, HWND hWnd, const RGNDATA *dirtyRegion) final {
-        return GetD3D9()->Present(src, dst, hWnd, dirtyRegion, 0);
-    }
-    
-    HRESULT STDMETHODCALLTYPE GetBackBuffer(UINT BackBuffer, D3DBACKBUFFER_TYPE Type, IDirect3DSurface8** ppBackBuffer) final {
-      if (unlikely(ppBackBuffer == nullptr))
-        return D3DERR_INVALIDCALL;
+    HRESULT STDMETHODCALLTYPE Present(const RECT *src, const RECT *dst, HWND hWnd, const RGNDATA *dirtyRegion) final;
 
-      // Same logic as in D3D8Device::GetBackBuffer
-      if (BackBuffer >= m_backBuffers.size() || m_backBuffers[BackBuffer] == nullptr) {
-        Com<d3d9::IDirect3DSurface9> pSurface9;
-        HRESULT res = GetD3D9()->GetBackBuffer(BackBuffer, (d3d9::D3DBACKBUFFER_TYPE)Type, &pSurface9);
-
-        if (likely(SUCCEEDED(res))) {
-          m_backBuffers[BackBuffer] = new D3D8Surface(GetParent(), std::move(pSurface9));
-          *ppBackBuffer = m_backBuffers[BackBuffer].ref();
-        }
-
-        return res;
-      }
-
-      *ppBackBuffer = m_backBuffers[BackBuffer].ref();
-      return D3D_OK;
-    }
+    HRESULT STDMETHODCALLTYPE GetBackBuffer(
+        UINT                BackBuffer,
+        D3DBACKBUFFER_TYPE  Type,
+        IDirect3DSurface8** ppBackBuffer) final;
 
   private:
+
     std::vector<Com<D3D8Surface, false>> m_backBuffers;
 
   };
